01-metadata/industrial-monitor/requirements.txt
pymodbus
asyncua
rich
plotext
pyyaml